+++ NEWS.md
@@ -6,6 +6,8 @@
 - New method `$clone()` for `LazyFrame` (#347).
 - `$with_column()` is now deprecated (following upstream `polars`). It will be
   removed in 0.9.0. It should be replaced with `$with_columns()` (#313).
+- New lazy function translated: `concat_str()` to concatenate several columns
+  into one (#349).
 
 # polars 0.7.0